requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
pydantic==2.6.0
pydantic_core==2.16.1
PyMySQL==1.1.0
python-dotenv==1.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
